+++ src/norm/pragmas.nim
@@ -37,13 +37,13 @@ template check*(val: string) {.pragma.}
   ## Add a ``CHECK <CONDITION>`` constraint.
 
 template unique* {.pragma.}
-## Add a ``UNIQUE`` constraint.
+  ## Add a ``UNIQUE`` constraint.
 
 template onDelete*(val: string) {.pragma.}
   ## Add an ``ON DELETE <POLICY>`` constraint.
 
 template onUpdate*(val: string) {.pragma.}
-## Add an ``ON UPDATE <POLICY>`` constraint.
+  ## Add an ``ON UPDATE <POLICY>`` constraint.
 
 template table*(val: string) {.pragma.}
   ## Set table name. Lowercased type name is used when unset.