+++ src/norm/sqlite.nim
@@ -7,8 +7,7 @@ SQLite Backend
 ]##
 
 
-import strutils, macros
-import typetraits
+import strutils, macros, typetraits
 import db_sqlite
 
 import chronicles
@@ -39,7 +38,7 @@ proc getTable*(T: type): string =
   ]##
 
   when T.hasCustomPragma(table): T.getCustomPragmaVal(table)
-  else: T.name.toLowerAscii()
+  else: ($T).toLowerAscii()
 
 proc getDbType(fieldRepr: FieldRepr): string =
   ## SQLite-specific mapping from Nim types to SQL data types.
@@ -111,33 +110,32 @@ proc genDropTableQueries*(dbObjReprs: seq[ObjRepr]): seq[string] =
 proc genInsertQuery*(obj: object, force: bool): SqlQuery =
   ## Generate ``INSERT`` query for an object.
 
-  var fields: seq[string]
-
-  for field, _ in obj.fieldPairs:
-    if force or not obj[field].hasCustomPragma(ro):
-      fields.add field
+  let
+    fields = obj.fieldNames(force)
+    placeholders = '?'.repeat(fields.len)
 
   result = sql "INSERT INTO $# ($#) VALUES ($#)" % [type(obj).getTable(), fields.join(", "),
-                                                    '?'.repeat(fields.len).join(", ")]
+                                                    placeholders.join(", ")]
 
 proc genGetOneQuery*(obj: object): SqlQuery =
   ## Generate ``SELECT`` query to fetch a single record for an object.
 
-  sql "SELECT $# FROM $# WHERE id = ?" % [obj.fieldNames.join(", "), type(obj).getTable()]
+  sql "SELECT $# FROM $# WHERE id = ?" % [obj.fieldNames(force=true).join(", "),
+                                          type(obj).getTable()]
 
 proc genGetManyQuery*(obj: object): SqlQuery =
   ## Generate ``SELECT`` query to fetch multiple records for an object.
 
-  sql "SELECT $# FROM $# LIMIT ? OFFSET ?" % [obj.fieldNames.join(", "), type(obj).getTable()]
+  sql "SELECT $# FROM $# LIMIT ? OFFSET ?" % [obj.fieldNames(force=true).join(", "),
+                                              type(obj).getTable()]
 
 proc getUpdateQuery*(obj: object, force: bool): SqlQuery =
   ## Generate ``UPDATE`` query for an object.
 
   var fieldsWithPlaceholders: seq[string]
 
-  for field, value in obj.fieldPairs:
-    if force or not obj[field].hasCustomPragma(ro):
-      fieldsWithPlaceholders.add field & " = ?"
+  for field in obj.fieldNames(force):
+    fieldsWithPlaceholders.add field & " = ?"
 
   result = sql "UPDATE $# SET $# WHERE id = ?" % [type(obj).getTable(),
                                                   fieldsWithPlaceholders.join(", ")]
@@ -152,7 +150,7 @@ template genWithDb(connection, user, password, database: string,
                         dbOthers: NimNode): untyped {.dirty.} =
   ## Generate ``withDb`` template.
 
-  template withDb(body: untyped): untyped {.dirty.} =
+  template withDb*(body: untyped): untyped {.dirty.} =
     ##[ A wrapper for actions that require DB connection. Defines CRUD procs to work with the DB,
     as well as ``createTables`` and ``dropTables`` procs.
 
@@ -190,13 +188,7 @@ template genWithDb(connection, user, password, database: string,
         By default, readonly fields are not inserted. Use ``force=true`` to insert all fields.
         ]##
 
-        var values: seq[string]
-
-        for _, value in obj.fieldPairs:
-          when force or not obj[_].hasCustomPragma(ro):
-            values.add $value
-
-        obj.id = dbConn.insertID(genInsertQuery(obj, force), values).int
+        obj.id = dbConn.insertID(genInsertQuery(obj, force), obj.toRow(force)).int
 
       template getOne(obj: var object, id: int) =
         ## Read a record from DB and store it into an existing object instance.
@@ -239,13 +231,7 @@ template genWithDb(connection, user, password, database: string,
         By default, readonly fields are not updated. Use ``force=true`` to update all fields.
         ]##
 
-        var values: seq[string]
-
-        for _, value in obj.fieldPairs:
-          when force or not obj[_].hasCustomPragma(ro):
-            values.add $value
-
-        dbConn.exec(getUpdateQuery(obj, force), values & $obj.id)
+        dbConn.exec(getUpdateQuery(obj, force), obj.toRow(force) & $obj.id)
 
       template delete(obj: var object) =
         ## Delete a record in DB by object's id. The id is set to 0 after the deletion.
@@ -290,37 +276,6 @@ macro db*(connection, user, password, database: string, body: untyped): untyped 
 
   The macro generates ``withDb`` template that wraps all DB interations.
   ]##
-
-  runnableExamples:
-    import db_sqlite
-
-    db(":memory:", "", "", ""):
-      type
-        User {.table: "users".} = object
-          email: string
-          age: int
-        Book {.table: "books".} = object
-          title: string
-          author {.fk: User.}: string
-
-      ## Define custom DB procs:
-      proc getUsersByEmail(email: string): seq[User] =
-        dbConn.getAllRows(sql "SELECT id, email, age FROM users WHERE email = ?", email).to User
-
-    withDb:
-      createTables()
-
-      ## Instantiate an object and insert it as a record:
-      var user = User(email: "hello@norm.nim", age: 30)
-      user.insert()
-
-      ## Retrieve the newly created record as an object:
-      doAssert User.getOne(user.id).email == "hello@norm.nim"
-
-      ## Use custom DB proc defined in ``db`` block:
-      doAssert getUsersByEmail("hello@norm.nim") == @[user]
-
-      dropTables()
 
   result = newStmtList()